Replace ContainerScope.CustomScope with ContainerTests.CustomScope.
